+++ src/Event/Listener/EventListenerConfigException.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Comquer\Event\EventListener;
+namespace Comquer\Event\Listener;
 
 use RuntimeException;
