+++ test/Event/EventDispatcherTest.php
@@ -6,9 +6,9 @@ use Comquer\DomainIntegration\Event\EventQueue;
 use Comquer\DomainIntegration\Event\EventQueueItemPublisher;
 use Comquer\DomainIntegration\Event\EventStore;
 use Comquer\Event\EventDispatcher;
-use Comquer\Event\EventQueueItem;
-use Comquer\Event\EventSubscription;
-use Comquer\Event\EventSubscriptionProvider;
+use Comquer\Event\Queue\EventQueueItem;
+use Comquer\Event\Subscription\EventSubscription;
+use Comquer\Event\Subscription\EventSubscriptionProvider;
 use ComquerTest\Fixture\Event\ItemAdded;
 use ComquerTest\Fixture\Event\UpdateShoppingListProjection;
 use PHPUnit\Framework\TestCase;
